Extract `ViewMailer`
