+++ src/MailerInterface.php
@@ -28,26 +28,6 @@ use Throwable;
 interface MailerInterface
 {
     /**
-     * Creates a new message instance and optionally composes its body content via view rendering.
-     *
-     * @param array<string, string>|string|null $view the view to be used for rendering the message body.
-     * This can be:
-     * - a string, which represents the view name for rendering the HTML body of the email.
-     *   In this case, the text body will be generated by applying `strip_tags()` to the HTML body.
-     * - an array with 'html' and/or 'text' elements. The 'html' element refers to the view name
-     *   for rendering the HTML body, while 'text' element is for rendering the text body. For example,
-     *   `['html' => 'contact-html', 'text' => 'contact-text']`.
-     * - null, meaning the message instance will be returned without body content.
-     * @param array $viewParameters The parameters (name-value pairs)
-     * that will be extracted and available in the view file.
-     * @param array $layoutParameters The parameters (name-value pairs)
-     * that will be extracted and available in the layout file.
-     *
-     * @return MessageInterface The message instance.
-     */
-    public function compose($view = null, array $viewParameters = [], array $layoutParameters = []): MessageInterface;
-
-    /**
      * Sends the given email message.
      *
      * @param MessageInterface $message The email message instance to be sent.
@@ -67,28 +47,4 @@ interface MailerInterface
      * @return SendResults The result object that contains all messages and errors for failed sent messages.
      */
     public function sendMultiple(array $messages): SendResults;
-
-    /**
-     * Returns a new instance with the specified message body template.
-     *
-     * @param MessageBodyTemplate $template The message body template instance.
-     *
-     * This method MUST be implemented in such a way as to retain the immutability of the mailer,
-     * and MUST return an instance that has the new message body template instance.
-     *
-     * @return self The new instance.
-     */
-    public function withTemplate(MessageBodyTemplate $template): self;
-
-    /**
-     * Returns a new instance with the specified locale.
-     *
-     * @param string $locale The locale code.
-     *
-     * This method MUST be implemented in such a way as to retain the immutability of the mailer,
-     * and MUST return an instance that has the new message body renderer instance.
-     *
-     * @return self The new instance.
-     */
-    public function withLocale(string $locale): self;
 }
