+++ tests/TestCase.php
@@ -10,8 +10,8 @@ use ReflectionClass;
 use Yiisoft\Files\FileHelper;
 use Yiisoft\Mailer\MailerInterface;
 use Yiisoft\Mailer\Message;
-use Yiisoft\Mailer\MessageBodyRenderer;
-use Yiisoft\Mailer\MessageBodyTemplate;
+use Yiisoft\Mailer\ViewMailer\MessageBodyRenderer;
+use Yiisoft\Mailer\ViewMailer\MessageBodyTemplate;
 use Yiisoft\Mailer\MessageInterface;
 use Yiisoft\Mailer\Tests\TestAsset\DummyMailer;
 use Yiisoft\Test\Support\Container\SimpleContainer;
